+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add contribution guidelines `CONTRIBUTING.md`
+- Support `LoadTextContent` function
 
 ## [0.3.1] - 2022-03-05